GET POST PUT DELETE de proveedor hecho

+++ src/Sucursal.js
@@ -33,14 +33,14 @@ class Sucursal {
   }
 
   validarSiEsDeSucursal(res, nombreSucursalRecibido) {
-   let esValido= false;
+    let esValido = false;
     let usuarioLogueado = this.obtenerUsuarioLogueado(res);
-    console.log(usuarioLogueado.sucursal+"Scursal del logueado")
-    console.log("sucursal revubido"+ nombreSucursalRecibido)
+    console.log(usuarioLogueado.sucursal + "Scursal del logueado")
+    console.log("sucursal revubido" + nombreSucursalRecibido)
     if (!(nombreSucursalRecibido == usuarioLogueado.sucursal)) {
       throw new ErrorDeIncidencia("Entrando a sucursal que no corresponde")
     }
-    esValido= true;
+    esValido = true;
     return esValido;
   }
 
@@ -183,12 +183,15 @@ class Sucursal {
   }
 
   async agregarProveedor(res, prov) {
-    let proveedorBuscado = await this.buscarProveedorPorCuil(prov.cuilProveedor);
+    let proveedorBuscado = await this.hayProveedor(prov.cuilProveedor);
     if (proveedorBuscado) {
       throw new Error("Ya existe un proveedor con ese Cuil");
     }
-    console.log("error Importante");
     await prov.save();
+  }
+
+  async hayProveedor(cuil) {
+    return await Proveedor.findOne({ cuilProveedor: cuil });
   }
 
   async agregarNotificacion(res, nuevaNotificacion) {
@@ -252,7 +255,7 @@ class Sucursal {
       await prod.save();
       return true;
     }
-    
+
   }
   async listaDeMovimientos() {
     return await Movimiento.find().lean();
@@ -281,6 +284,10 @@ class Sucursal {
   }
 
   async eliminarProveedor(id) {
+    let unProveedor = await this.buscarProveedorPorId(id);
+    if (!unProveedor) {
+      throw new Error("El Id recibido no existe!")
+    }
     await Proveedor.findByIdAndDelete(id);
   }
 
@@ -302,9 +309,13 @@ class Sucursal {
     return await Empleado.findById(id).lean();
   }
   async buscarProveedorPorCuil(cuil) {
-    console.log("recibiendo cuil en metodo buscar "+cuil)
-    return await Proveedor.findOne({ cuilProveedor: cuil })
+    let unProveedor = await Proveedor.findOne({ cuilProveedor: cuil });
+    if (!unProveedor) {
+      throw new Error("Proveedor no existe")
+    }
+    return unProveedor
   }
+
   async buscarUsuarioPorLegajo(legajo) {
     return await Empleado.find({ legajo: legajo });
   }
@@ -364,7 +375,7 @@ class Sucursal {
 
   async editarProveedor(id, params) {
     let unProveedor = await this.buscarProveedorPorCuil(params.cuilProveedor);
-    if (unProveedor && id != unProveedor._id ) {
+    if (unProveedor && id != unProveedor._id) {
       throw new Error("El numero de cuil ya se encuentra asignado a otro Proveedor!");
     }
     return await Proveedor.findByIdAndUpdate(id, params);
